requirements.txt
easyocr
pdf2image
streamlit
transformers
faiss-cpu
sentence-transformers
gtts
torch